requirement.txt
discord.py==2.3.2
httpx==0.27.0
python-dotenv==1.0.0